requirements.txt
# Core dependencies
streamlit>=1.22.0
streamlit-drawable-canvas>=0.8.0
pillow>=9.0.0
numpy>=1.22.0
pandas>=1.4.0
pydicom>=2.3.0

# DICOM utilities
pylibjpeg
pylibjpeg-libjpeg

# Image processing
scikit-image>=0.19.0
opencv-python-headless>=4.5.0

# Translation support
deep-translator>=1.10.0

# API interactions
requests>=2.27.0

# PDF generation
fpdf2>=2.5.0

# Optional: UMLS utilities (smaller dependency)
umls-api-client>=0.0.6